refactor(abstract-utxo): split sign.ts into separate modules
Split the large sign.ts file into smaller, more focused modules:
- signLegacyTransaction.ts for legacy transaction format
- signPsbt.ts for PSBT-related signing functionality
- SigningError.ts for error classes

This refactoring improves code organization and makes it easier to maintain
the different signing pathways.

Issue: BTC-2806
