+++ modules/abstract-utxo/src/transaction/fixedScript/index.ts
@@ -3,6 +3,8 @@ export { explainPsbtWasm } from './explainPsbtWasm';
 export { parseTransaction } from './parseTransaction';
 export { CustomChangeOptions } from './parseOutput';
 export { verifyTransaction } from './verifyTransaction';
-export { signTransaction, Musig2Participant } from './signTransaction';
-export * from './sign';
+export { signTransaction } from './signTransaction';
+export { Musig2Participant } from './signPsbt';
+export * from './signLegacyTransaction';
+export * from './SigningError';
 export * from './replayProtection';
